requirements.txt
anthropic>=0.25.0
python-dotenv
